refactor: rename mapToRecordFilterOperand file to match function name
Made-with: Cursor
