+++ packages/twenty-front/src/modules/object-record/record-filter/utils/mapRLSOperandToRecordFilterOperand.ts
